# Conflicts:
#	cl_dll/ev_hldm.h
#	cl_dll/hud.h
#	common/Platform.h
#	common/const.h
#	dlls/cdll_dll.h
#	dlls/client.cpp
#	dlls/multiplay_gamerules.cpp
#	dlls/player.h
#	dlls/scriptevent.h
#	dlls/weapons.h
#	linux/Makefile.hl_cdll
#	projects/vs2019/hl_cdll.vcxproj

+++ cl_dll/hl/hl_weapons.cpp
@@ -17,7 +17,6 @@
 #include "cbase.h"
 #include "monsters.h"
 #include "weapons.h"
-#include "nodes.h"
 #include "player.h"
 #include "weapons/CGrapple.h"
 #include "weapons/CEagle.h"
@@ -271,11 +270,11 @@ CBaseEntity::FireBulletsPlayer
 Only produces random numbers to match the server ones.
 =====================
 */
-Vector CBaseEntity::FireBulletsPlayer ( ULONG cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker, int shared_rand )
+Vector CBaseEntity::FireBulletsPlayer (unsigned int cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker, int shared_rand )
 {
 	float x, y, z;
 
-	for ( ULONG iShot = 1; iShot <= cShots; iShot++ )
+	for (unsigned int iShot = 1; iShot <= cShots; iShot++ )
 	{
 		if ( pevAttacker == NULL )
 		{
